Starting a cpanfile for my most commonly-used modules.
Make it easy to get a new box up and running with the stuff I expect.

I might release it to CPAN as a Task::BeLike::*, but I tend to check out my
dotfiles repo on each box I'm going to do any serious work on anyway, so just
throwing this through cpanm is probably enough, without having to spam CPAN with
stuff that's only really for my personal use.