Write backups to ~/tmp/vimbackups
Write a backup snapshot of the file each time I save it from vim.

This would have saved me some arse-ache at least once recently.

Yes, version control is the real way forward, but for e.g. quickly editing a
config file, or hacking on a file that's not yet committed and accidentally
forgetting it was open in another editor and overwriting it, it's worth the disk
space.  I can always set up a cron job to delete backups older than a day or
two.