Perl version of cherry-pick-from-branch
The simple shell function I used for selecting the commits on a feature branch
which aren't in master and cherry-picking them to another branch to create a PR
for hotfix into a release branch, turned into Perl and made more flexible and
safer.

TODO: if it hits a merge conflict or a cherry pick fails for any other reason,
you're stuck.  Maybe it should dump out a list of the remaining SHAs - maybe
even as a list of git cherry-pick $sha commands, which you could then run after
fixing it?

TODO: I added some logic to skip any commits we already have (for instance,
because we bailed part way through on hitting a conflict) - but it compares the
SHAs, and of course the cherry-picked commit will have a different SHA.  Hmm.