ci alias improvements - add GH num, auto-append, diff.
Check the branch name; if we can parse out a GitHub issue number from it
(e.g. it contains "gh-42" or similar) then prepend the commit message
we're about to edit with it (for easy compliance with a requirement at
my new employer that commits should reference the GH issue number).

Automatically drop into append mode in vim to save doing it manually -
particularly useful if we did prepend a GitHub issue number, as you'll
end up with the cursor after it ready to start typing.

Add a succinct git diff --stat to the output to see at a glance the
scale of the changes you're committing.