Two-space indent for Puppet config.
This could be done with exrc and dumping a .vimrc in the repo in
question, but I don't want to risk accidentally opening some other repo
which has a malicious .vimrc in it.

So, just match on files being opened from that path and set the options
I want here instead.