Quick script for easy git branch switching
Quite often I hop between branches, and this makes it easier.

Run it from within your git checkout, as e.g. `gsb pattern`, and it will
look for branches which match `pattern`.  If only one is found (as you
may expect if you provided e.g. an issue number), it'll switch to it
immediately.  If it matched multiple branches, it will present a list of
matches, and ask you to enter the number of the one you want.

Why?  It's quicker and move convenient than moving my hand off to the
mouse, or first searching to find the branch name then typing it (even
with tab completion).