Avoid "invalid ioctl for device" on X logins
The stty call to disable terminal flow control has caused an error on X
login for ages, but I've never fixed it because my machines tend to stay
up, so it's a rare annoyance that I always think "one day I'll fix
that".

It appears that, finally, today may be That Day.