pipeid utility script to identify pipe endpoints
Added a quick and dirty script to identify the endpoints of pipes.

For e.g. strace might tell me it's waiting on a pipe:

```
[davidp@slideaway:~]$ sudo strace -s255 -T -y -yy -C -p
5537
strace: Process 5537 attached
select(9, [6<pipe:[529293520]> 8<pipe:[529293521]>], NULL, NULL, NULL
```

So, to see what is at the end of one of those pipes,

```
[davidp@slideaway:~]$ pipeid 529293520
Looking for endpoints for pipe:[529293520]
PIPE[529293520]: 5537 (/usr/sbin/fcgiwrap), 11602 (/usr/bin/perl)
```

It's a bit ugly, but it automates chaining a couple of commands I always have
to stop and think about.