Remove PERL5OPT=-M5.010
This was set for laziness, but it causes occasional problems (for e.g.
test suites which fail with it set), and doesn't really save me all that
much.  I can write use 5.012 or whatever version I want in code (and in
fact should, so that that code will work the same for other people as it
does for me!), and one-liners, which IIRC were the main reason I set
this, can just use -E.