Set ripgrep conf env var if we have a conf file
ripgrep doesn't look for config files in any particular location, you
set this env var to point it at a conf file if you want to.

I may refine this in future, possibly adding an `rg` function which
first looks in your current directory and walks upwards looking for a
ripgrep config, falling back to the one in your homedir, then executes
ripgrep - so that you can add project-specific ones and expect them to
work.

(Maybe have it be smart enough that, if there is a project-specific one
found and also one in $HOME/.ripgreprc, cat them both to a temp file
(project-specific last, so it overrides your general one) and point
ripgrep at that.)

For now, though, the most common thing I set in the config is some globs
to exclude particular dirs, which are not going to clash.