sshretry resolve host via config; show wait time
Support resolving the given hostname via ssh config - for example if we
were called as `sshretry myalias`, and `.ssh/config` contained:

  Host myalias
    Hostname realhostname.example.com

... then it's no good pinging myalias to see if it's up, because that
isn't going to resolve, so we need to ping the Hostname that the config
says.

Also, time how long it took and output that - sometimes useful to be
able to see how long the box was unavailable for.