Easier "around" notation in getlines
Easier way to get a set of lines around a given line for context.

Also, doco.

Wanted to use Term::ANSIColor to color the line asked for and dim the
context, but needs a bit of playing around so that it doesn't spew noise
if used in a pipe rather than STDOUT to a terminal - looks like I could
use Term::ANSIColor::Conditional to fix that, but it's not in core.

May deal with that at some point.  For now, I just get a prefixed arrow.