Have setup script create auto-update cronjob
Check if we already have a cronjob to update automatically, if not, Make
It So.

Also, avoid error symlinking Zim config if the Zim config dir doesn't
exist - I have Zim on workstations, but many boxes my dotfiles are on
are not GUI workstations so will obviously not have it.