Full path to aws-cli binary
My cron path on some boxes at least doesn't include /usr/local/bin so
I'm getting errors from my auto-update cronjob when it sources the
profile afterwards.

Maybe a better fix would be for the ./setup script to only source the
possibly-updated profile at the end if we're in an interactive session,
though.  (Or fix up $PATH in crontab, but I don't really want to have to
do that on every box.)