Record gsb branch switch history
Start recording `gsb` branch switches to a history file.

In time, I will implement a history option to have it list the last few
branches you switched from/to and let you pick the one you need, but for
that to be useful, I'll need to be collecting the data already.

It's also handy to have if looking back on times to see when you were on
certain branches etc.