Add my handy `ssh-as` command
Useful `ssh-as` command I've been using for some time now - looks up
which server a given installation is on, and SSHes to it, using sudo to
become that install's user (unless you use `-m`/`--me` to stay as me).

Defaults to the live server for them, unless `-d` / `--dev` / `--qa` is
used.