Simple utility for joining values from lines
A simple utility to read lines of input from STDIN, optionally split
them by a chosen regex, then output them joined by a customisable string
(defaults to a comma) optionally using a sprintf format to format them.

Useful to be able to, for instance, pipe it a list of IDs, one per line
from a file, and turn them in to comma-separated values ready to paste
in to an SQL query's `WHERE id IN ( ... )`.

Or to take a list of values and use sprintf formatting to zero-pad /
round etc.

Future ideas:
* if splitting lines, allow picking only some columns to output
* be able to specify leading and trailing text which is output before
  and after all the values e.g. --before 'WHERE id IN('  --after ')'